Add support for deserializing LIST responses
The HubSpot list responses are different kind of beasts so they require
different handling. Purely generic stuff has been implemented to deal with
parsing the root level "companies" or "contacts" or what ever HubSpot
throws into the root level JSON object.

It's not depending on the names of the props in the DTO, but rather IList types
and generic arguments that are decendandts of IHubSpotEntity so it should handle
anything.