Implement company search by domain
CompanyListRequestOption is no loger a HubSpotEntity as
HubSpot has decided that this exact endpoint is just a
standard JSON serialization instead of the name value mapping.

Added another List operation to the HubSpotBaseClient which
POSTs instead so that a payload can be passed along.

Added functional test of the search by email which is a company
search by domain.